pyproject.toml
[project]
name = "demo-fastapi"
version = "0.1.0"
description = "brief fastapi demo"
authors = [
    {name = "thai le"}
]
readme = "README.md"
requires-python = "3.13.1"
dependencies = [
    "annotated-types (==0.7.0)",
    "anyio (==4.8.0)",
    "cachetools (==5.5.1)",
    "certifi (==2025.1.31)",
    "charset-normalizer (==3.4.1)",
    "click (==8.1.8)",
    "dnspython (==2.7.0)",
    "email-validator (==2.2.0)",
    "fastapi (==0.115.8)",
    "fastapi-cli (==0.0.7)",
    "google-ai-generativelanguage (==0.6.15)",
    "google-api-core (==2.24.1)",
    "google-api-python-client (==2.160.0)",
    "google-auth (==2.38.0)",
    "google-auth-httplib2 (==0.2.0)",
    "google-generativeai (==0.8.4)",
    "googleapis-common-protos (==1.66.0)",
    "grpcio (==1.70.0)",
    "grpcio-status (==1.70.0)",
    "h11 (==0.14.0)",
    "httpcore (==1.0.7)",
    "httplib2 (==0.22.0)",
    "httptools (==0.6.4)",
    "httpx (==0.28.1)",
    "idna (==3.10)",
    "jinja2 (==3.1.5)",
    "markdown-it-py (==3.0.0)",
    "markupsafe (==3.0.2)",
    "mdurl (==0.1.2)",
    "proto-plus (==1.26.0)",
    "protobuf (==5.29.3)",
    "pyasn1 (==0.6.1)",
    "pyasn1-modules (==0.4.1)",
    "pydantic (==2.10.6)",
    "pydantic-settings (==2.7.1)",
    "pydantic-core (==2.27.2)",
    "pygments (==2.19.1)",
    "pyparsing (==3.2.1)",
    "python-dotenv (==1.0.1)",
    "python-multipart (==0.0.20)",
    "pyyaml (==6.0.2)",
    "requests (==2.32.3)",
    "rich (==13.9.4)",
    "rich-toolkit (==0.13.2)",
    "rsa (==4.9)",
    "shellingham (==1.5.4)",
    "sniffio (==1.3.1)",
    "starlette (==0.45.3)",
    "tqdm (==4.67.1)",
    "typer (==0.15.1)",
    "typing-extensions (==4.12.2)",
    "uritemplate (==4.1.1)",
    "urllib3 (==2.3.0)",
    "uvicorn (==0.34.0)",
    "uvloop (==0.21.0)",
    "watchfiles (==1.0.4)",
    "websockets (==14.2)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "firebase-admin (>=6.6.0,<7.0.0)",
    "sentence-transformers (>=3.4.1,<4.0.0)",
    "qdrant-client (>=1.13.2,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
